Tweaked logging Refactored imports
